Factor out escalate fn
