+++ src/lib.rs
@@ -2,6 +2,7 @@ pub mod admin;
 pub mod endpoint;
 pub mod systemd_api;
 pub mod types;
+pub mod utils;
 
 pub mod pb {
     pub mod admin {
